backend/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.8.0

# LLM Provider SDKs
openai>=1.0.0
anthropic>=0.18.0
google-genai>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Security
python-multipart>=0.0.6
